Fix Error on Resizing

+++ src/main/image_loader.ts
@@ -81,7 +81,9 @@ export async function generateResizedCache(
 		);
 		if (files.length === 0) return;
 
-		await !fs.rm(resizedCacheDir, { recursive: true, force: true });
+		if (!fsSync.existsSync(resizedCacheDir)) {
+			await !fs.rm(resizedCacheDir, { recursive: true, force: true });
+		}
 		await fs.mkdir(resizedCacheDir, { recursive: true });
 
 		for (const input of files) {
@@ -94,7 +96,7 @@ export async function generateResizedCache(
 				.toFile(outputPath);
 		}
 	} catch (error) {
-		console.error(`Error Recreating "${resizedCacheDir}"`, error);
+		console.error("Error on creating resized images:", error);
 		throw error;
 	}
 }
